chore(mimo): Commit initial helm chart template
There's a few things that aren't completed in this template yet, in all
likelihood most, if not all of what's in the secret_provider will need
to be moved to just plain environment variables supplied with a config
map as most of these don't seem like they would have any secrets in
them (likely that the rp pulls them in itself?).

It's not clear what settings are needed fo rthe service account just
yet, and deploying with pipeline variables isn't entirely clear yet
either.