fips: stop checking system libraries
We run this fips-detect tool in the builder image for our container
image builds, and we deploy the output image. Checking the filesystem of
the builder image for shared libraries has no bearing on anything we
actually deploy, and we do not care one way or the other what is going
on there.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>